commit bpm changes

+++ platform/component/bpm/source/src/main/java/com/sinosoft/one/bpm/util/ProcessInstanceBOCache.java
@@ -94,10 +94,11 @@ public class ProcessInstanceBOCache {
 		if(StringUtils.isNotBlank(businessId)) {
 			processInstanceIdCache.remove(processId + "_" + businessId);
 			businessIdCache.remove(processInstanceId);
+			
+			ProcessInstanceBOInfo info = processInstanceBOService.getProcessInstanceBOInfo(processInstanceId);
+			if(info != null) {
+				processInstanceBOService.removeProcessInstanceBOInfo(info);
+			}
 		}
-		
-		ProcessInstanceBOInfo info = processInstanceBOService.getProcessInstanceBOInfo(processInstanceId);
-		processInstanceBOService.removeProcessInstanceBOInfo(info);
 	}
-	
 }
